optionally extract ddb column name and type from struct tags
StructToAttributeMap and AttributeMapToStruct will look for a "ddb:"
struct tag.

If the tag is not present for the field, the field will serialize to an
"S"-typed column named the same as the struct field.

If the tag is present, the converter will look for exactly 1 or 2
comma-delimited fields in the tag.
The first part is the DDB column name. The second part is the column
type.

If the entire tag is "-", the field is ignored.
If the column name is "-", the default column name is used instead.
If the tag type is "", the default is used instead.
If the tag type is neither "S" nor "N", it is an error and the field is
ignored.